provider/azure: rework to new stack API
