--- go/src/koding/kites/kloud/provider/azure/glide.yaml
@@ -1,3 +1,0 @@
-package: github.com/koding/kloud-provider-azure
-import:
-- package: github.com/Azure/azure-sdk-for-go
